Changed to categretial problem

+++ DARNN/data_class.py
@@ -25,13 +25,13 @@ class StockDataset(Dataset):
                               device = self.device)
         self.y = pd_data[ylabel].copy().fillna(0).values
         self.y = torch.tensor(self.y, 
-                              dtype = torch.uint8,
+                              dtype = torch.float64,
                               requires_grad=False,
                               device = self.device)
         self.target = pd_data[ylabel, 'RET'].copy().fillna(0).values
-        self.target = (self.target > 0).astype('uint8')
+        self.target = (self.target > 0).astype('float64')
         self.target = torch.tensor(self.target, 
-                                   dtype = torch.uint8,
+                                   dtype = torch.float64,
                                    requires_grad=False,
                                    device = self.device)
         self.total_time = len(pd_data)
